// src/models/enums/health-choice-item.enum.ts
export enum HealthChoiceItem {
  YES_NO = 'yes_no',
  YES_NO_MAYBE = 'yes_no_maybe',
  HEALTH_GOAL = 'health_goal',
  HEALTH_CONDITION = 'health_condition',
  PHYSICAL_ACTIVITY_LEVEL = 'physical_activity_level',
  DIETARY_PREFERENCE = 'dietary_preference',
  SYMPTOMS = 'symptoms',
  GENDER = 'gender',
  FITNESS_FOCUS_AREA = 'fitness_focus_area',
  MOOD = 'mood',
  FEELING_SCALE = 'feeling_scale',
  CONCERNS_OR_SYMPTOMS = 'concerns_or_symptoms',
  HEALTH_CHANGES = 'health_changes',
  UNUSUAL_SENSATION = 'unusual_sensation',
  MENTAL_STRESS = 'mental_stress',
  SLEEP_PATTERN = 'sleep_pattern',
  LIMITATIONS = 'limitations',
  MENSTRUAL_CYCLE_REGULARITY = 'menstrual_cycle_irregularity',
  MENSTRUAL_CYCLE_DISCOMFORT = 'menstrual_cycle_discomfort',
  REPRODUCTIVE_HEALTH_DISORDERS = 'reproductive_health_disorders',
  FAVORITE_FOOD = 'favorite_food',
}